small code cleanup Prevent a potential NPE remove some not needed errors fix smart folders in the SAS warning tab for find similar add a new tab for team todos

+++ src/main/java/com/swissas/quickfix/ConvertToTeamQuickfix.java
@@ -24,9 +24,11 @@ public class ConvertToTeamQuickfix implements LocalQuickFix {
 	@Override
 	public void applyFix(@NotNull Project project, @NotNull ProblemDescriptor descriptor) {
 		PsiDocTag tag = (PsiDocTag)descriptor.getPsiElement();
+		if(tag != null) {
 		PsiDocTag newAuthorTag = JavaPsiFacade.getElementFactory(project).createDocTagFromText(
 				"@author " + SwissAsStorage.getInstance().getMyTeam());
-		tag.getParent().addBefore(newAuthorTag, tag);
-		tag.delete();
+			tag.getParent().addBefore(newAuthorTag, tag);
+			tag.delete();
+		}
 	}
 }
